update: 在庫チェックページのロジック分離、コンポーネント分割
