+++ src/presentations/components/InventoryCheckPage/hooks/index.js
@@ -1,0 +1,1 @@
+export * from "./useInventoryCheck";
